Split auto updater and settings manager into libraries
